clementine_core

Module aggregator

Structs

Aggregator

Aggregator struct. This struct is responsible for aggregating partial signatures from the verifiers. It will have in total 3 * num_operator + 1 aggregated nonces. [0] -> Aggregated nonce for the move transaction. [1..num_operator + 1] -> Aggregated nonces for the operator_takes transactions. [num_operator + 1..2 * num_operator + 1] -> Aggregated nonces for the slash_or_take transactions. [2 * num_operator + 1..3 * num_operator + 1] -> Aggregated nonces for the burn transactions. For now, we do not have the last bit.

AggregatorServer

Aggregator server wrapper that manages background tasks.

OperatorId

Wrapper struct that renders the operator id in the logs.

ParticipatingOperators

Wrapper struct that matches operator clients with their ids.

ParticipatingVerifiers

Wrapper struct that matches verifier clients with their ids.

VerifierId

Wrapper struct that renders the verifier id in the logs.

Enums

EntityId