clementine_core
Module bitcoin_syncer

Bitcoin Syncer

This module provides common utilities to fetch Bitcoin state. Other modules can use this module to operate over Bitcoin.

Structs

BitcoinSyncer

BitcoinSyncerTask

A task that syncs Bitcoin blocks from the Bitcoin node to the local database.

FinalizedBlockFetcherTask

Enums

BitcoinSyncerEvent

Events emitted by the Bitcoin syncer. It emits the block_id of the block in the db that was saved.

Traits

BlockHandler

Trait for handling new blocks as they are finalized

Functions

set_initial_block_info_if_not_exists

If no block info exists in the DB, fetches the current block from the RPC and initializes the DB.