clementine_core::builder::sighash
Function create_operator_sighash_stream

pub fn create_operator_sighash_stream(
    db: Database,
    operator_xonly_pk: XOnlyPublicKey,
    config: BridgeConfig,
    deposit_data: DepositData,
    deposit_blockhash: BlockHash,
) -> impl Stream<Item = Result<(TapSighash, SignatureInfo), BridgeError>>

These operator sighashes are needed so that each operator can share the signatures with each verifier, so that verifiers have the ability to burn the burn connector of operators. WIP: Update if the design changes. This function generates Kickoff Timeout TX, Already Disproved TX, and Disprove TX for each sequential_collateral_tx and kickoff_utxo. It yields the sighashes for these tx’s for the input that has operators burn connector. Possible future optimization: Each verifier already generates some of these TX’s in create_operator_sighash_stream() It is possible to for verifiers somehow return the required sighashes for operator signatures there too. But operators only needs to use sighashes included in this function.