clementine_core::builder::transaction
Module challenge

Functions

create_challenge_timeout_txhandler

Creates a TxHandler for the no challenge. This transaction used when no one sends a challenge tx, so that operator can spend kickoff finalizer to finalize the kickoff.

create_challenge_txhandler

Creates a TxHandler for the challenge. This transaction is for covering the operators’ cost for a challenge to prevent people from maliciously challenging them and causing them to lose money.

create_disprove_txhandler

Creates a TxHandler for the disprove_tx. This transaction will be sent by NofN, meaning that the operator was malicious. This transaction burns the operator’s burn connector, kicking the operator out of the system.

create_operator_challenge_ack_txhandler

Creates a TxHandler for the operator_challenge_ACK_tx. This transaction will is used so that the operator can acknowledge the challenge and reveal the preimage for the corresponding watchtower. If the operator does not reveal the preimage, the NofN will be able to spend the output after 0.5 week using operator_challenge_NACK_tx.

create_operator_challenge_nack_txhandler

Creates a TxHandler for the operator_challenge_NACK_tx. This transaction will force the operator to reveal the preimage for the corresponding watchtower since if they do not reveal the preimage, the NofN will be able to spend the output after 0.5 week, which will prevent the operator from sending assert_begin_tx.

create_watchtower_challenge_timeout_txhandler

Creates the watchtower challenge timeout txhandler. This tx needs to be sent by operators when a watchtower doesn’t send a challenge, otherwise operator will be forced to reveal their preimage.

create_watchtower_challenge_txhandler

Creates a TxHandler for the watchtower_challenge_tx. This transaction is sent by the watchtowers to reveal their Groth16 proofs with their public inputs for the longest chain proof. The data is encoded as 32 byte script pubkeys of taproot utxos, and a single at max 80 byte OP_RETURN utxo.