clementine_core::builder::transaction
Function create_emergency_stop_txhandler

pub fn create_emergency_stop_txhandler(
    deposit_data: &mut DepositData,
    move_to_vault_txhandler: &TxHandler,
    paramset: &'static ProtocolParamset,
) -> Result<TxHandler<Unsigned>, BridgeError>

Creates a TxHandler for the emergency_stop_tx.

This transaction moves funds to the address controlled by the security council from the move-to-vault txout. Used to stop the deposit in case of a security issue. The moved funds will eventually be redeposited using the replacement deposit tx.

Arguments

deposit_data - Mutable reference to the deposit data for the transaction.

move_to_vault_txhandler - Reference to the move-to-vault transaction handler.

paramset - Protocol parameter set.

Returns

A TxHandler for the emergency stop transaction, or a BridgeError if construction fails.