clementine_core::builder::transaction
Function create_round_txhandlers

pub fn create_round_txhandlers(
    paramset: &'static ProtocolParamset,
    round_idx: RoundIndex,
    operator_data: &OperatorData,
    kickoff_winternitz_keys: &KickoffWinternitzKeys,
    prev_ready_to_reimburse: Option<&TxHandler>,
) -> Result<Vec<TxHandler>, BridgeError>

Creates the round and ready-to-reimburse txhandlers for a specific operator and round index. These transactions currently include round tx, ready to reimburse tx, and unspent kickoff txs.

Arguments

paramset - Protocol parameter set.

round_idx - The index of the round.

operator_data - Data for the operator.

kickoff_winternitz_keys - All winternitz keys of the operator.

prev_ready_to_reimburse - Previous ready-to-reimburse txhandler, if any, to not create the full collateral chain if we already have the previous round’s ready to reimburse txhandler.

Returns

A vector of TxHandler for the round, ready-to-reimburse, and unspent kickoff transactions, or a BridgeError if construction fails.