clementine_core::builder::transaction

Module operator_collateral

Collaterals

This module contains the logic for creating the round_tx, ready_to_reimburse_tx, and unspent_kickoff_tx transactions. These transactions are used to control the sequence of transactions in the withdrawal process and limits the number of withdrawals the operator can make in a given time period.

The flow is as follows: round_tx -> ready_to_reimburse_tx -> round_tx -> ...

The round_tx is used to create a collateral for the withdrawal, kickoff utxos for the current round and the reimburse connectors for the previous round.

Enums

RoundTxInput

Functions

create_assert_timeout_txhandlers

Creates a vector of TxHandler for assert_timeout_tx transactions.

create_burn_unused_kickoff_connectors_txhandler

Creates a TxHandler for burning unused kickoff connectors.

create_ready_to_reimburse_txhandler

Creates a TxHandler for the ready_to_reimburse_tx.

create_round_nth_txhandler

Creates the nth (1-indexed) round_txhandler and reimburse_generator_txhandler pair for a specific operator.

create_round_txhandler

Creates a TxHandler for round_tx.

create_unspent_kickoff_txhandlers

Creates a vector of TxHandler for unspent_kickoff_tx transactions. These transactions can be sent if an operator sends ReadyToReimburse transaction without spending all the kickoff utxos of the round.