clementine_core::citrea

Module BRIDGE_CONTRACT

Generated by the following Solidity interface…

library Bridge {
    struct MerkleProof {
        bytes intermediateNodes;
        uint256 blockHeight;
        uint256 index;
    }
    struct Transaction {
        bytes4 version;
        bytes2 flag;
        bytes vin;
        bytes vout;
        bytes witness;
        bytes4 locktime;
    }
    struct UTXO {
        bytes32 txId;
        bytes4 outputId;
    }
}

interface BRIDGE_CONTRACT {
    error InvalidInitialization();
    error NotInitializing();
    error OwnableInvalidOwner(address owner);
    error OwnableUnauthorizedAccount(address account);

    event Deposit(bytes32 wtxId, bytes32 txId, address recipient, uint256 timestamp, uint256 depositId);
    event DepositReplaced(uint256 index, bytes32 oldTxId, bytes32 newTxId);
    event DepositScriptUpdate(bytes depositPrefix, bytes depositSuffix);
    event DepositTransferFailed(bytes32 wtxId, bytes32 txId, address recipient, uint256 timestamp, uint256 depositId);
    event FailedDepositVaultUpdated(address oldVault, address newVault);
    event Initialized(uint64 version);
    event OperatorUpdated(address oldOperator, address newOperator);
    event OwnershipTransferStarted(address indexed previousOwner, address indexed newOwner);
    event OwnershipTransferred(address indexed previousOwner, address indexed newOwner);
    event ReplaceScriptUpdate(bytes replacePrefix, bytes replaceSuffix);
    event SafeWithdrawal(Bridge.Transaction payoutTx, Bridge.UTXO spentUtxo, uint256 index);
    event Withdrawal(Bridge.UTXO utxo, uint256 index, uint256 timestamp);

    function CODESEP_POS() external view returns (bytes memory);
    function EPOCH() external view returns (bytes memory);
    function INPUT_INDEX() external view returns (bytes memory);
    function KEY_VERSION() external view returns (bytes memory);
    function LIGHT_CLIENT() external view returns (address);
    function SCHNORR_VERIFIER_PRECOMPILE() external view returns (address);
    function SIGHASH_ALL_HASH_TYPE() external view returns (bytes memory);
    function SIGHASH_SINGLE_ANYONECANPAY_HASH_TYPE() external view returns (bytes memory);
    function SPEND_TYPE_EXT() external view returns (bytes memory);
    function SPEND_TYPE_NO_EXT() external view returns (bytes memory);
    function SYSTEM_CALLER() external view returns (address);
    function acceptOwnership() external;
    function batchWithdraw(bytes32[] memory txIds, bytes4[] memory outputIds) external payable;
    function deposit(Bridge.Transaction memory moveTx, Bridge.MerkleProof memory proof, bytes32 shaScriptPubkeys) external;
    function depositAmount() external view returns (uint256);
    function depositPrefix() external view returns (bytes memory);
    function depositSuffix() external view returns (bytes memory);
    function depositTxIds(uint256) external view returns (bytes32);
    function failedDepositVault() external view returns (address);
    function getAggregatedKey() external view returns (bytes memory);
    function getWithdrawalCount() external view returns (uint256);
    function initialize(bytes memory _depositPrefix, bytes memory _depositSuffix, uint256 _depositAmount) external;
    function initialized() external view returns (bool);
    function operator() external view returns (address);
    function owner() external view returns (address);
    function pendingOwner() external view returns (address);
    function processedTxIds(bytes32) external view returns (bool);
    function renounceOwnership() external;
    function replaceDeposit(Bridge.Transaction memory replaceTx, Bridge.MerkleProof memory proof, uint256 idToReplace, bytes32 shaScriptPubkeys) external;
    function replacePrefix() external view returns (bytes memory);
    function replaceSuffix() external view returns (bytes memory);
    function safeWithdraw(Bridge.Transaction memory prepareTx, Bridge.MerkleProof memory prepareProof, Bridge.Transaction memory payoutTx, bytes memory blockHeader, bytes memory scriptPubKey) external payable;
    function setDepositScript(bytes memory _depositPrefix, bytes memory _depositSuffix) external;
    function setFailedDepositVault(address _failedDepositVault) external;
    function setOperator(address _operator) external;
    function setReplaceScript(bytes memory _replacePrefix, bytes memory _replaceSuffix) external;
    function transferOwnership(address newOwner) external;
    function withdraw(bytes32 txId, bytes4 outputId) external payable;
    function withdrawalUTXOs(uint256) external view returns (bytes32 txId, bytes4 outputId);
}

…which was generated by the following JSON ABI:

[
  {
    "type": "function",
    "name": "CODESEP_POS",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "bytes",
        "internalType": "bytes"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "EPOCH",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "bytes",
        "internalType": "bytes"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "INPUT_INDEX",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "bytes",
        "internalType": "bytes"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "KEY_VERSION",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "bytes",
        "internalType": "bytes"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "LIGHT_CLIENT",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "address",
        "internalType": "contract BitcoinLightClient"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "SCHNORR_VERIFIER_PRECOMPILE",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "address",
        "internalType": "address"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "SIGHASH_ALL_HASH_TYPE",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "bytes",
        "internalType": "bytes"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "SIGHASH_SINGLE_ANYONECANPAY_HASH_TYPE",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "bytes",
        "internalType": "bytes"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "SPEND_TYPE_EXT",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "bytes",
        "internalType": "bytes"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "SPEND_TYPE_NO_EXT",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "bytes",
        "internalType": "bytes"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "SYSTEM_CALLER",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "address",
        "internalType": "address"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "acceptOwnership",
    "inputs": [],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "batchWithdraw",
    "inputs": [
      {
        "name": "txIds",
        "type": "bytes32[]",
        "internalType": "bytes32[]"
      },
      {
        "name": "outputIds",
        "type": "bytes4[]",
        "internalType": "bytes4[]"
      }
    ],
    "outputs": [],
    "stateMutability": "payable"
  },
  {
    "type": "function",
    "name": "deposit",
    "inputs": [
      {
        "name": "moveTx",
        "type": "tuple",
        "internalType": "struct Bridge.Transaction",
        "components": [
          {
            "name": "version",
            "type": "bytes4",
            "internalType": "bytes4"
          },
          {
            "name": "flag",
            "type": "bytes2",
            "internalType": "bytes2"
          },
          {
            "name": "vin",
            "type": "bytes",
            "internalType": "bytes"
          },
          {
            "name": "vout",
            "type": "bytes",
            "internalType": "bytes"
          },
          {
            "name": "witness",
            "type": "bytes",
            "internalType": "bytes"
          },
          {
            "name": "locktime",
            "type": "bytes4",
            "internalType": "bytes4"
          }
        ]
      },
      {
        "name": "proof",
        "type": "tuple",
        "internalType": "struct Bridge.MerkleProof",
        "components": [
          {
            "name": "intermediateNodes",
            "type": "bytes",
            "internalType": "bytes"
          },
          {
            "name": "blockHeight",
            "type": "uint256",
            "internalType": "uint256"
          },
          {
            "name": "index",
            "type": "uint256",
            "internalType": "uint256"
          }
        ]
      },
      {
        "name": "shaScriptPubkeys",
        "type": "bytes32",
        "internalType": "bytes32"
      }
    ],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "depositAmount",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "uint256",
        "internalType": "uint256"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "depositPrefix",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "bytes",
        "internalType": "bytes"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "depositSuffix",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "bytes",
        "internalType": "bytes"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "depositTxIds",
    "inputs": [
      {
        "name": "",
        "type": "uint256",
        "internalType": "uint256"
      }
    ],
    "outputs": [
      {
        "name": "",
        "type": "bytes32",
        "internalType": "bytes32"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "failedDepositVault",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "address",
        "internalType": "address"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "getAggregatedKey",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "bytes",
        "internalType": "bytes"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "getWithdrawalCount",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "uint256",
        "internalType": "uint256"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "initialize",
    "inputs": [
      {
        "name": "_depositPrefix",
        "type": "bytes",
        "internalType": "bytes"
      },
      {
        "name": "_depositSuffix",
        "type": "bytes",
        "internalType": "bytes"
      },
      {
        "name": "_depositAmount",
        "type": "uint256",
        "internalType": "uint256"
      }
    ],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "initialized",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "bool",
        "internalType": "bool"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "operator",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "address",
        "internalType": "address"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "owner",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "address",
        "internalType": "address"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "pendingOwner",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "address",
        "internalType": "address"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "processedTxIds",
    "inputs": [
      {
        "name": "",
        "type": "bytes32",
        "internalType": "bytes32"
      }
    ],
    "outputs": [
      {
        "name": "",
        "type": "bool",
        "internalType": "bool"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "renounceOwnership",
    "inputs": [],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "replaceDeposit",
    "inputs": [
      {
        "name": "replaceTx",
        "type": "tuple",
        "internalType": "struct Bridge.Transaction",
        "components": [
          {
            "name": "version",
            "type": "bytes4",
            "internalType": "bytes4"
          },
          {
            "name": "flag",
            "type": "bytes2",
            "internalType": "bytes2"
          },
          {
            "name": "vin",
            "type": "bytes",
            "internalType": "bytes"
          },
          {
            "name": "vout",
            "type": "bytes",
            "internalType": "bytes"
          },
          {
            "name": "witness",
            "type": "bytes",
            "internalType": "bytes"
          },
          {
            "name": "locktime",
            "type": "bytes4",
            "internalType": "bytes4"
          }
        ]
      },
      {
        "name": "proof",
        "type": "tuple",
        "internalType": "struct Bridge.MerkleProof",
        "components": [
          {
            "name": "intermediateNodes",
            "type": "bytes",
            "internalType": "bytes"
          },
          {
            "name": "blockHeight",
            "type": "uint256",
            "internalType": "uint256"
          },
          {
            "name": "index",
            "type": "uint256",
            "internalType": "uint256"
          }
        ]
      },
      {
        "name": "idToReplace",
        "type": "uint256",
        "internalType": "uint256"
      },
      {
        "name": "shaScriptPubkeys",
        "type": "bytes32",
        "internalType": "bytes32"
      }
    ],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "replacePrefix",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "bytes",
        "internalType": "bytes"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "replaceSuffix",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "bytes",
        "internalType": "bytes"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "safeWithdraw",
    "inputs": [
      {
        "name": "prepareTx",
        "type": "tuple",
        "internalType": "struct Bridge.Transaction",
        "components": [
          {
            "name": "version",
            "type": "bytes4",
            "internalType": "bytes4"
          },
          {
            "name": "flag",
            "type": "bytes2",
            "internalType": "bytes2"
          },
          {
            "name": "vin",
            "type": "bytes",
            "internalType": "bytes"
          },
          {
            "name": "vout",
            "type": "bytes",
            "internalType": "bytes"
          },
          {
            "name": "witness",
            "type": "bytes",
            "internalType": "bytes"
          },
          {
            "name": "locktime",
            "type": "bytes4",
            "internalType": "bytes4"
          }
        ]
      },
      {
        "name": "prepareProof",
        "type": "tuple",
        "internalType": "struct Bridge.MerkleProof",
        "components": [
          {
            "name": "intermediateNodes",
            "type": "bytes",
            "internalType": "bytes"
          },
          {
            "name": "blockHeight",
            "type": "uint256",
            "internalType": "uint256"
          },
          {
            "name": "index",
            "type": "uint256",
            "internalType": "uint256"
          }
        ]
      },
      {
        "name": "payoutTx",
        "type": "tuple",
        "internalType": "struct Bridge.Transaction",
        "components": [
          {
            "name": "version",
            "type": "bytes4",
            "internalType": "bytes4"
          },
          {
            "name": "flag",
            "type": "bytes2",
            "internalType": "bytes2"
          },
          {
            "name": "vin",
            "type": "bytes",
            "internalType": "bytes"
          },
          {
            "name": "vout",
            "type": "bytes",
            "internalType": "bytes"
          },
          {
            "name": "witness",
            "type": "bytes",
            "internalType": "bytes"
          },
          {
            "name": "locktime",
            "type": "bytes4",
            "internalType": "bytes4"
          }
        ]
      },
      {
        "name": "blockHeader",
        "type": "bytes",
        "internalType": "bytes"
      },
      {
        "name": "scriptPubKey",
        "type": "bytes",
        "internalType": "bytes"
      }
    ],
    "outputs": [],
    "stateMutability": "payable"
  },
  {
    "type": "function",
    "name": "setDepositScript",
    "inputs": [
      {
        "name": "_depositPrefix",
        "type": "bytes",
        "internalType": "bytes"
      },
      {
        "name": "_depositSuffix",
        "type": "bytes",
        "internalType": "bytes"
      }
    ],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "setFailedDepositVault",
    "inputs": [
      {
        "name": "_failedDepositVault",
        "type": "address",
        "internalType": "address"
      }
    ],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "setOperator",
    "inputs": [
      {
        "name": "_operator",
        "type": "address",
        "internalType": "address"
      }
    ],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "setReplaceScript",
    "inputs": [
      {
        "name": "_replacePrefix",
        "type": "bytes",
        "internalType": "bytes"
      },
      {
        "name": "_replaceSuffix",
        "type": "bytes",
        "internalType": "bytes"
      }
    ],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "transferOwnership",
    "inputs": [
      {
        "name": "newOwner",
        "type": "address",
        "internalType": "address"
      }
    ],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "withdraw",
    "inputs": [
      {
        "name": "txId",
        "type": "bytes32",
        "internalType": "bytes32"
      },
      {
        "name": "outputId",
        "type": "bytes4",
        "internalType": "bytes4"
      }
    ],
    "outputs": [],
    "stateMutability": "payable"
  },
  {
    "type": "function",
    "name": "withdrawalUTXOs",
    "inputs": [
      {
        "name": "",
        "type": "uint256",
        "internalType": "uint256"
      }
    ],
    "outputs": [
      {
        "name": "txId",
        "type": "bytes32",
        "internalType": "bytes32"
      },
      {
        "name": "outputId",
        "type": "bytes4",
        "internalType": "bytes4"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "event",
    "name": "Deposit",
    "inputs": [
      {
        "name": "wtxId",
        "type": "bytes32",
        "indexed": false,
        "internalType": "bytes32"
      },
      {
        "name": "txId",
        "type": "bytes32",
        "indexed": false,
        "internalType": "bytes32"
      },
      {
        "name": "recipient",
        "type": "address",
        "indexed": false,
        "internalType": "address"
      },
      {
        "name": "timestamp",
        "type": "uint256",
        "indexed": false,
        "internalType": "uint256"
      },
      {
        "name": "depositId",
        "type": "uint256",
        "indexed": false,
        "internalType": "uint256"
      }
    ],
    "anonymous": false
  },
  {
    "type": "event",
    "name": "DepositReplaced",
    "inputs": [
      {
        "name": "index",
        "type": "uint256",
        "indexed": false,
        "internalType": "uint256"
      },
      {
        "name": "oldTxId",
        "type": "bytes32",
        "indexed": false,
        "internalType": "bytes32"
      },
      {
        "name": "newTxId",
        "type": "bytes32",
        "indexed": false,
        "internalType": "bytes32"
      }
    ],
    "anonymous": false
  },
  {
    "type": "event",
    "name": "DepositScriptUpdate",
    "inputs": [
      {
        "name": "depositPrefix",
        "type": "bytes",
        "indexed": false,
        "internalType": "bytes"
      },
      {
        "name": "depositSuffix",
        "type": "bytes",
        "indexed": false,
        "internalType": "bytes"
      }
    ],
    "anonymous": false
  },
  {
    "type": "event",
    "name": "DepositTransferFailed",
    "inputs": [
      {
        "name": "wtxId",
        "type": "bytes32",
        "indexed": false,
        "internalType": "bytes32"
      },
      {
        "name": "txId",
        "type": "bytes32",
        "indexed": false,
        "internalType": "bytes32"
      },
      {
        "name": "recipient",
        "type": "address",
        "indexed": false,
        "internalType": "address"
      },
      {
        "name": "timestamp",
        "type": "uint256",
        "indexed": false,
        "internalType": "uint256"
      },
      {
        "name": "depositId",
        "type": "uint256",
        "indexed": false,
        "internalType": "uint256"
      }
    ],
    "anonymous": false
  },
  {
    "type": "event",
    "name": "FailedDepositVaultUpdated",
    "inputs": [
      {
        "name": "oldVault",
        "type": "address",
        "indexed": false,
        "internalType": "address"
      },
      {
        "name": "newVault",
        "type": "address",
        "indexed": false,
        "internalType": "address"
      }
    ],
    "anonymous": false
  },
  {
    "type": "event",
    "name": "Initialized",
    "inputs": [
      {
        "name": "version",
        "type": "uint64",
        "indexed": false,
        "internalType": "uint64"
      }
    ],
    "anonymous": false
  },
  {
    "type": "event",
    "name": "OperatorUpdated",
    "inputs": [
      {
        "name": "oldOperator",
        "type": "address",
        "indexed": false,
        "internalType": "address"
      },
      {
        "name": "newOperator",
        "type": "address",
        "indexed": false,
        "internalType": "address"
      }
    ],
    "anonymous": false
  },
  {
    "type": "event",
    "name": "OwnershipTransferStarted",
    "inputs": [
      {
        "name": "previousOwner",
        "type": "address",
        "indexed": true,
        "internalType": "address"
      },
      {
        "name": "newOwner",
        "type": "address",
        "indexed": true,
        "internalType": "address"
      }
    ],
    "anonymous": false
  },
  {
    "type": "event",
    "name": "OwnershipTransferred",
    "inputs": [
      {
        "name": "previousOwner",
        "type": "address",
        "indexed": true,
        "internalType": "address"
      },
      {
        "name": "newOwner",
        "type": "address",
        "indexed": true,
        "internalType": "address"
      }
    ],
    "anonymous": false
  },
  {
    "type": "event",
    "name": "ReplaceScriptUpdate",
    "inputs": [
      {
        "name": "replacePrefix",
        "type": "bytes",
        "indexed": false,
        "internalType": "bytes"
      },
      {
        "name": "replaceSuffix",
        "type": "bytes",
        "indexed": false,
        "internalType": "bytes"
      }
    ],
    "anonymous": false
  },
  {
    "type": "event",
    "name": "SafeWithdrawal",
    "inputs": [
      {
        "name": "payoutTx",
        "type": "tuple",
        "indexed": false,
        "internalType": "struct Bridge.Transaction",
        "components": [
          {
            "name": "version",
            "type": "bytes4",
            "internalType": "bytes4"
          },
          {
            "name": "flag",
            "type": "bytes2",
            "internalType": "bytes2"
          },
          {
            "name": "vin",
            "type": "bytes",
            "internalType": "bytes"
          },
          {
            "name": "vout",
            "type": "bytes",
            "internalType": "bytes"
          },
          {
            "name": "witness",
            "type": "bytes",
            "internalType": "bytes"
          },
          {
            "name": "locktime",
            "type": "bytes4",
            "internalType": "bytes4"
          }
        ]
      },
      {
        "name": "spentUtxo",
        "type": "tuple",
        "indexed": false,
        "internalType": "struct Bridge.UTXO",
        "components": [
          {
            "name": "txId",
            "type": "bytes32",
            "internalType": "bytes32"
          },
          {
            "name": "outputId",
            "type": "bytes4",
            "internalType": "bytes4"
          }
        ]
      },
      {
        "name": "index",
        "type": "uint256",
        "indexed": false,
        "internalType": "uint256"
      }
    ],
    "anonymous": false
  },
  {
    "type": "event",
    "name": "Withdrawal",
    "inputs": [
      {
        "name": "utxo",
        "type": "tuple",
        "indexed": false,
        "internalType": "struct Bridge.UTXO",
        "components": [
          {
            "name": "txId",
            "type": "bytes32",
            "internalType": "bytes32"
          },
          {
            "name": "outputId",
            "type": "bytes4",
            "internalType": "bytes4"
          }
        ]
      },
      {
        "name": "index",
        "type": "uint256",
        "indexed": false,
        "internalType": "uint256"
      },
      {
        "name": "timestamp",
        "type": "uint256",
        "indexed": false,
        "internalType": "uint256"
      }
    ],
    "anonymous": false
  },
  {
    "type": "error",
    "name": "InvalidInitialization",
    "inputs": []
  },
  {
    "type": "error",
    "name": "NotInitializing",
    "inputs": []
  },
  {
    "type": "error",
    "name": "OwnableInvalidOwner",
    "inputs": [
      {
        "name": "owner",
        "type": "address",
        "internalType": "address"
      }
    ]
  },
  {
    "type": "error",
    "name": "OwnableUnauthorizedAccount",
    "inputs": [
      {
        "name": "account",
        "type": "address",
        "internalType": "address"
      }
    ]
  }
]

Structs

BRIDGE_CONTRACTInstance

A BRIDGE_CONTRACT instance.

CODESEP_POSCall

Function with signature CODESEP_POS() and selector 0xb2497e70.

CODESEP_POSReturn

Container type for the return parameters of the CODESEP_POS() function.

Deposit

Event with signature Deposit(bytes32,bytes32,address,uint256,uint256) and selector 0xa82453ca34121b3ecb910d957824e27c5dc6465315949facd15fb72886490058.

DepositReplaced

Event with signature DepositReplaced(uint256,bytes32,bytes32) and selector 0x4d7c644a48da4c7857af62a00bad9806f0388564f22955ed846d938c244047f0.

DepositScriptUpdate

Event with signature DepositScriptUpdate(bytes,bytes) and selector 0x80bd1fdfe157286ce420ee763f91748455b249605748e5df12dad9844402bafc.

DepositTransferFailed

Event with signature DepositTransferFailed(bytes32,bytes32,address,uint256,uint256) and selector 0xabd361bc68da04a386a8de9d0fb3044cca0856cbd86e9e4a63237e015b3e4bb9.

EPOCHCall

Function with signature EPOCH() and selector 0xa0dc2758.

EPOCHReturn

Container type for the return parameters of the EPOCH() function.

FailedDepositVaultUpdated

Event with signature FailedDepositVaultUpdated(address,address) and selector 0x79250b96878fd457364d1c1b77a660973c4f4ab67bda5e2fdb42caaa4d515f9d.

INPUT_INDEXCall

Function with signature INPUT_INDEX() and selector 0x428bcd35.

INPUT_INDEXReturn

Container type for the return parameters of the INPUT_INDEX() function.

Initialized

Event with signature Initialized(uint64) and selector 0xc7f505b2f371ae2175ee4913f4499e1f2633a7b5936321eed1cdaeb6115181d2.

InvalidInitialization

Custom error with signature InvalidInitialization() and selector 0xf92ee8a9.

KEY_VERSIONCall

Function with signature KEY_VERSION() and selector 0x1369ac3e.

KEY_VERSIONReturn

Container type for the return parameters of the KEY_VERSION() function.

LIGHT_CLIENTCall

Function with signature LIGHT_CLIENT() and selector 0xe613ae00.

LIGHT_CLIENTReturn

Container type for the return parameters of the LIGHT_CLIENT() function.

NotInitializing

Custom error with signature NotInitializing() and selector 0xd7e6bcf8.

OperatorUpdated

Event with signature OperatorUpdated(address,address) and selector 0xfbe5b6cbafb274f445d7fed869dc77a838d8243a22c460de156560e8857cad03.

OwnableInvalidOwner

Custom error with signature OwnableInvalidOwner(address) and selector 0x1e4fbdf7.

OwnableUnauthorizedAccount

Custom error with signature OwnableUnauthorizedAccount(address) and selector 0x118cdaa7.

OwnershipTransferStarted

Event with signature OwnershipTransferStarted(address,address) and selector 0x38d16b8cac22d99fc7c124b9cd0de2d3fa1faef420bfe791d8c362d765e22700.

OwnershipTransferred

Event with signature OwnershipTransferred(address,address) and selector 0x8be0079c531659141344cd1fd0a4f28419497f9722a3daafe3b4186f6b6457e0.

ReplaceScriptUpdate

Event with signature ReplaceScriptUpdate(bytes,bytes) and selector 0x6c9ac69a5e351d3e7ac9be95040d29a264d1ce6a409ca9f042c64c66c3f2a23a.

SCHNORR_VERIFIER_PRECOMPILECall

Function with signature SCHNORR_VERIFIER_PRECOMPILE() and selector 0x9a4f308d.

SCHNORR_VERIFIER_PRECOMPILEReturn

Container type for the return parameters of the SCHNORR_VERIFIER_PRECOMPILE() function.

SIGHASH_ALL_HASH_TYPECall

Function with signature SIGHASH_ALL_HASH_TYPE() and selector 0x0bd89ab7.

SIGHASH_ALL_HASH_TYPEReturn

Container type for the return parameters of the SIGHASH_ALL_HASH_TYPE() function.

SIGHASH_SINGLE_ANYONECANPAY_HASH_TYPECall

Function with signature SIGHASH_SINGLE_ANYONECANPAY_HASH_TYPE() and selector 0x6cf7d641.

SIGHASH_SINGLE_ANYONECANPAY_HASH_TYPEReturn

Container type for the return parameters of the SIGHASH_SINGLE_ANYONECANPAY_HASH_TYPE() function.

SPEND_TYPE_EXTCall

Function with signature SPEND_TYPE_EXT() and selector 0x23dacd29.

SPEND_TYPE_EXTReturn

Container type for the return parameters of the SPEND_TYPE_EXT() function.

SPEND_TYPE_NO_EXTCall

Function with signature SPEND_TYPE_NO_EXT() and selector 0x092ac5d4.

SPEND_TYPE_NO_EXTReturn

Container type for the return parameters of the SPEND_TYPE_NO_EXT() function.

SYSTEM_CALLERCall

Function with signature SYSTEM_CALLER() and selector 0xd761753e.

SYSTEM_CALLERReturn

Container type for the return parameters of the SYSTEM_CALLER() function.

SafeWithdrawal

Event with signature SafeWithdrawal((bytes4,bytes2,bytes,bytes,bytes,bytes4),(bytes32,bytes4),uint256) and selector 0xd77102e5369b5b1a9db1972cb3de26ee79abc69de5cde41eeaa67fe3939c1c55.

Withdrawal

Event with signature Withdrawal((bytes32,bytes4),uint256,uint256) and selector 0x3311a04a346a103ac115cca33028a2bc82f1964805860d0d3fc84a2772496ada.

acceptOwnershipCall

Function with signature acceptOwnership() and selector 0x79ba5097.

acceptOwnershipReturn

Container type for the return parameters of the acceptOwnership() function.

batchWithdrawCall

Function with signature batchWithdraw(bytes32[],bytes4[]) and selector 0x19854623.

batchWithdrawReturn

Container type for the return parameters of the batchWithdraw(bytes32[],bytes4[]) function.

depositAmountCall

Function with signature depositAmount() and selector 0x419759f5.

depositAmountReturn

Container type for the return parameters of the depositAmount() function.

depositCall

Function with signature deposit((bytes4,bytes2,bytes,bytes,bytes,bytes4),(bytes,uint256,uint256),bytes32) and selector 0xfb11d7b9.

depositPrefixCall

Function with signature depositPrefix() and selector 0x5b4f894d.

depositPrefixReturn

Container type for the return parameters of the depositPrefix() function.

depositReturn

Container type for the return parameters of the deposit((bytes4,bytes2,bytes,bytes,bytes,bytes4),(bytes,uint256,uint256),bytes32) function.

depositSuffixCall

Function with signature depositSuffix() and selector 0x8752b6b2.

depositSuffixReturn

Container type for the return parameters of the depositSuffix() function.

depositTxIdsCall

Function with signature depositTxIds(uint256) and selector 0x06592167.

depositTxIdsReturn

Container type for the return parameters of the depositTxIds(uint256) function.

failedDepositVaultCall

Function with signature failedDepositVault() and selector 0xa670e7ed.

failedDepositVaultReturn

Container type for the return parameters of the failedDepositVault() function.

getAggregatedKeyCall

Function with signature getAggregatedKey() and selector 0x3c918b6c.

getAggregatedKeyReturn

Container type for the return parameters of the getAggregatedKey() function.

getWithdrawalCountCall

Function with signature getWithdrawalCount() and selector 0x781952a8.

getWithdrawalCountReturn

Container type for the return parameters of the getWithdrawalCount() function.

initializeCall

Function with signature initialize(bytes,bytes,uint256) and selector 0x41260137.

initializeReturn

Container type for the return parameters of the initialize(bytes,bytes,uint256) function.

initializedCall

Function with signature initialized() and selector 0x158ef93e.

initializedReturn

Container type for the return parameters of the initialized() function.

operatorCall

Function with signature operator() and selector 0x570ca735.

operatorReturn

Container type for the return parameters of the operator() function.

ownerCall

Function with signature owner() and selector 0x8da5cb5b.

ownerReturn

Container type for the return parameters of the owner() function.

pendingOwnerCall

Function with signature pendingOwner() and selector 0xe30c3978.

pendingOwnerReturn

Container type for the return parameters of the pendingOwner() function.

processedTxIdsCall

Function with signature processedTxIds(bytes32) and selector 0x4379caa5.

processedTxIdsReturn

Container type for the return parameters of the processedTxIds(bytes32) function.

renounceOwnershipCall

Function with signature renounceOwnership() and selector 0x715018a6.

renounceOwnershipReturn

Container type for the return parameters of the renounceOwnership() function.

replaceDepositCall

Function with signature replaceDeposit((bytes4,bytes2,bytes,bytes,bytes,bytes4),(bytes,uint256,uint256),uint256,bytes32) and selector 0x2594f107.

replaceDepositReturn

Container type for the return parameters of the replaceDeposit((bytes4,bytes2,bytes,bytes,bytes,bytes4),(bytes,uint256,uint256),uint256,bytes32) function.

replacePrefixCall

Function with signature replacePrefix() and selector 0x6b0b5a94.

replacePrefixReturn

Container type for the return parameters of the replacePrefix() function.

replaceSuffixCall

Function with signature replaceSuffix() and selector 0xf42cb4fc.

replaceSuffixReturn

Container type for the return parameters of the replaceSuffix() function.

safeWithdrawCall

Function with signature safeWithdraw((bytes4,bytes2,bytes,bytes,bytes,bytes4),(bytes,uint256,uint256),(bytes4,bytes2,bytes,bytes,bytes,bytes4),bytes,bytes) and selector 0x9072f747.

safeWithdrawReturn

Container type for the return parameters of the safeWithdraw((bytes4,bytes2,bytes,bytes,bytes,bytes4),(bytes,uint256,uint256),(bytes4,bytes2,bytes,bytes,bytes,bytes4),bytes,bytes) function.

setDepositScriptCall

Function with signature setDepositScript(bytes,bytes) and selector 0xf8e655d2.

setDepositScriptReturn

Container type for the return parameters of the setDepositScript(bytes,bytes) function.

setFailedDepositVaultCall

Function with signature setFailedDepositVault(address) and selector 0x85fb7151.

setFailedDepositVaultReturn

Container type for the return parameters of the setFailedDepositVault(address) function.

setOperatorCall

Function with signature setOperator(address) and selector 0xb3ab15fb.

setOperatorReturn

Container type for the return parameters of the setOperator(address) function.

setReplaceScriptCall

Function with signature setReplaceScript(bytes,bytes) and selector 0x7ec9732a.

setReplaceScriptReturn

Container type for the return parameters of the setReplaceScript(bytes,bytes) function.

transferOwnershipCall

Function with signature transferOwnership(address) and selector 0xf2fde38b.

transferOwnershipReturn

Container type for the return parameters of the transferOwnership(address) function.

withdrawCall

Function with signature withdraw(bytes32,bytes4) and selector 0x8786dba7.

withdrawReturn

Container type for the return parameters of the withdraw(bytes32,bytes4) function.

withdrawalUTXOsCall

Function with signature withdrawalUTXOs(uint256) and selector 0x471ba1e3.

withdrawalUTXOsReturn

Container type for the return parameters of the withdrawalUTXOs(uint256) function.

Enums

BRIDGE_CONTRACTCalls

Container for all the BRIDGE_CONTRACT function calls.

BRIDGE_CONTRACTErrors

Container for all the BRIDGE_CONTRACT custom errors.

BRIDGE_CONTRACTEvents

Container for all the BRIDGE_CONTRACT events.

Statics

BYTECODE

The creation / init bytecode of the contract.

DEPLOYED_BYTECODE

The runtime bytecode of the contract, as deployed on the network.

Functions

deploy

Deploys this contract using the given provider and constructor arguments, if any.

deploy_builder

Creates a RawCallBuilder for deploying this contract using the given provider and constructor arguments, if any.

new

Creates a new wrapper around an on-chain BRIDGE_CONTRACT contract instance.