clementine_core
Module cli

Command Line Interface

This module defines command line interface for server binaries. Clap is used for easy generation of help messages and handling arguments.

Structs

Args

Clementine (C) 2025 Chainway Limited

Enums

Actors

ConfigSource

Functions

get_cli_config

Gets configuration using CLI arguments, for binaries. If there are any errors, prints error and panics.

get_cli_config_from_args 🔒

Wrapped function for tests

get_config_source

Selects a configuration source for the main config or the protocol paramset.

parse_from 🔒

Parse given iterator. This is good for isolated environments, like tests.