clementine_core

Module constants

Modules

timeout 🔒

Structs

BURN_SCRIPT

Constants

DEFAULT_CHANNEL_SIZE

DEPOSIT_FINALIZATION_TIMEOUT

DEPOSIT_FINALIZE_STREAM_CREATION_TIMEOUT

ENTITY_COMP_DATA_POLL_TIMEOUT

ENTITY_STATUS_POLL_TIMEOUT

KEY_DISTRIBUTION_TIMEOUT

MAX_ALL_SESSIONS_BYTES

The maximum number of bytes that can be used by all nonce sessions. If it exceeds this limit, the verifier will delete the oldest nonce sessions. This limit is approximate, because it doesn’t take into account the internal extra bytes used in HashMap and VecDeque used in the AllSessions. It only takes into account bytes used for the secnonces.

MAX_BYTES_PER_WINTERNITZ_KEY

The maximum number of bytes per Winternitz key.

MAX_EXTRA_WATCHTOWERS

Maximum extra watchtowers that can be added to the deposit (in addition to verifiers). It is limited because each extra watchtower requires 2 additional utxos in the kickoff tx.

MAX_NUM_SESSIONS

The maximum number of nonce sessions that can be stored in the verifier. It is used so that the allsessions do not store too many small (1 nonce) sessions.

MAX_SCRIPT_REPLACEMENT_OPERATIONS

The maximum number of script replacement operations allowed in a single BitVM operation. This is a safeguard to prevent excessive resource usage and ensure that the BitVM protocol remains efficient and manageable. The limit is set to 100,000 operations, which is a reasonable upper bound for script replacement operations in the context of BitVM, which is normally a constant equal to 47544.

MAX_WINTERNITZ_DIGITS_PER_KEY

The maximum number of Winternitz digits per key. This is used to limit the size of the Winternitz public keys in the protocol to prevent excessive memory usage and ensure efficient processing. This value is achieved when signing a 32-byte message with a Winternitz key, resulting in a maximum of 64 + 4 digits per key, where the last 4 digits are used for the sum-check operation.

MIN_TAPROOT_AMOUNT

The minimum possible amount that a UTXO can have when created into a Taproot address.

NONCE_STREAM_CREATION_TIMEOUT

NON_EPHEMERAL_ANCHOR_AMOUNT

The amount of the non-ephemeral P2A anchor output.

NON_STANDARD_V3

NUM_NONCES_LIMIT

The maximum number of nonces that can be generated in a single nonce generation session. A single nonce takes 132 (musig2 secret nonce) bytes. We calculate NUM_NONCES so that a nonce session takes at maximum 150MB.

OPERATOR_GET_KEYS_TIMEOUT

OPERATOR_SIGS_STREAM_CREATION_TIMEOUT

OPERATOR_SIGS_TIMEOUT

OPTIMISTIC_PAYOUT_TIMEOUT

OVERALL_DEPOSIT_TIMEOUT

PARTIAL_SIG_STREAM_CREATION_TIMEOUT

PIPELINE_COMPLETION_TIMEOUT

PUBLIC_KEY_COLLECTION_TIMEOUT

RESTART_BACKGROUND_TASKS_TIMEOUT

SEND_OPERATOR_SIGS_TIMEOUT

SETUP_COMPLETION_TIMEOUT

TEN_MINUTES_IN_SECS

VERIFIER_SEND_KEYS_TIMEOUT

WITHDRAWAL_TIMEOUT