clementine_core::database

Module wrapper

Type Wrappers for Parsing

This module includes wrappers for easy parsing of the foreign types.

Macros

impl_bytea_wrapper_custom 🔒

Macro for implementing BYTEA-based SQL wrapper types with custom encoding/decoding

impl_bytea_wrapper_custom_with_error 🔒

Same as impl_bytea_wrapper_custom but with an encode function that returns a Result

impl_bytea_wrapper_default 🔒

Macro for implementing BYTEA-based SQL wrapper types using standard serialization

impl_text_wrapper_base 🔒

Macro to reduce boilerplate for [impl_text_wrapper_custom].

impl_text_wrapper_custom 🔒

Macro for implementing text-based SQL wrapper types with custom encoding/decoding

impl_text_wrapper_default 🔒

Macro for implementing text-based SQL wrapper types using standard string conversion

Structs

AddressDB

BlockHashDB

BlockHeaderDB

DepositParamsDB

EVMAddressDB

MessageDB

MusigAggNonceDB

MusigPubNonceDB

OutPointDB

PublicKeyDB

ReceiptDB

ScriptBufDB

SignatureDB

SignaturesDB

TxOutDB

TxidDB

UtxoDB

XOnlyPublicKeyDB