clementine_core
Module extended_bitcoin_rpc

Bitcoin Extended RPC Interface

Extended RPC interface communicates with the Bitcoin node. It features some common wrappers around typical RPC operations as well as direct communication interface with the Bitcoin node.

Tests

In tests, Bitcoind node and client are usually created using [crate::test::common::create_regtest_rpc]. Please refer to [crate::test::common] for using ExtendedBitcoinRpc in tests.

Structs

ExtendedBitcoinRpc

Bitcoin RPC wrapper. Extended RPC provides useful wrapper functions for common operations, as well as direct access to Bitcoin RPC.

RetryConfig

Enums

BitcoinRPCError

Errors that can occur during Bitcoin RPC operations.

Traits

RetryableError

Trait to determine if an error is retryable

Type Aliases

Result ๐