clementine_core::extended_bitcoin_rpc
Struct ExtendedBitcoinRpc

pub struct ExtendedBitcoinRpc {
    url: String,
    client: Arc<Client>,
    retry_config: RetryConfig,
}

Bitcoin RPC wrapper. Extended RPC provides useful wrapper functions for common operations, as well as direct access to Bitcoin RPC.

Fields
url: Stringclient: Arc<Client>retry_config: RetryConfig
Implementations

impl ExtendedBitcoinRpc

pub async fn connect( url: String, user: SecretString, password: SecretString, retry_config: Option<RetryConfig>, ) -> Result<Self, BitcoinRPCError>

Connects to Bitcoin RPC server with built-in retry mechanism.

This method attempts to connect to the Bitcoin RPC server and creates a new ExtendedBitcoinRpc instance. It includes retry logic that will retry connection attempts for retryable errors using exponential backoff.

Parameters

url - The RPC server URL

user - Username for RPC authentication

password - Password for RPC authentication

retry_config - Optional retry configuration. If None, uses default config.

Returns

Result<ExtendedBitcoinRpc>: A new ExtendedBitcoinRpc instance on success

Errors

BitcoinRPCError: If connection fails after all retry attempts or ping fails

pub async fn get_new_wallet_address(&self) -> Result<Address, BitcoinRPCError>

Generates a new Bitcoin address for the wallet.

pub async fn confirmation_blocks( &self, txid: &Txid, ) -> Result<u32, BitcoinRPCError>

Returns the number of confirmations for a transaction.

Parameters

txid: TXID of the transaction to check.

Returns

u32: The number of confirmations for the transaction.

Errors

BitcoinRPCError: If the transaction is not confirmed (0) or if there was an error retrieving the transaction info.

pub async fn get_current_chain_height(&self) -> Result<u32, BitcoinRPCError>

Retrieves the current blockchain height (number of blocks).

Returns

u32: Current block height

pub async fn collateral_check( &self, operator_data: &OperatorData, kickoff_wpks: &KickoffWinternitzKeys, paramset: &'static ProtocolParamset, ) -> Result<bool, BridgeError>

Checks if an operator’s collateral is valid and available for use.

This function validates the operator’s collateral by:

Verifying the collateral UTXO exists and has the correct amount

Creating the round transaction chain to track current collateral position

Determining if the current collateral UTXO in the chain is spent in a non-protocol tx, signaling the exit of operator from the protocol

Parameters

operator_data: Data about the operator including collateral funding outpoint

kickoff_wpks: Kickoff Winternitz public keys for round transaction creation

paramset: Protocol parameters

Returns

bool: true if the collateral is still usable, thus operator is still in protocol, false if the collateral is spent, thus operator is not in protocol anymore

Errors

BridgeError: If there was an error retrieving transaction data, creating round transactions, or checking UTXO status

pub async fn get_blockhash_of_tx( &self, txid: &Txid, ) -> Result<BlockHash, BitcoinRPCError>

Returns block hash of a transaction, if confirmed.

Parameters

txid: TXID of the transaction to check.

Returns

[bitcoin::BlockHash]: Block hash of the block that the transaction is in.

Errors

BitcoinRPCError: If the transaction is not confirmed (0) or if there was an error retrieving the transaction info.

pub async fn get_block_info_by_height( &self, height: u64, ) -> Result<(BlockHash, Header), BitcoinRPCError>

Retrieves the block header and hash for a given block height.

Arguments

height: Target block height.

Returns

([bitcoin::BlockHash], [bitcoin::block::Header]): A tuple containing the block hash and header.

pub async fn get_prevout_txs( &self, tx: &Transaction, ) -> Result<Vec<Transaction>, BitcoinRPCError>

Gets the transactions that created the inputs of a given transaction.

Arguments

tx - The transaction to get the previous transactions for

Returns

A vector of transactions that created the inputs of the given transaction.

pub async fn get_tx_of_txid( &self, txid: &Txid, ) -> Result<Transaction, BitcoinRPCError>

Gets the transaction data for a given transaction ID.

Parameters

txid: TXID of the transaction to check.

Returns

[bitcoin::Transaction]: Transaction itself.

pub async fn is_tx_on_chain(&self, txid: &Txid) -> Result<bool, BitcoinRPCError>

Checks if a transaction is on-chain.

Parameters

txid: TXID of the transaction to check.

Returns

bool: true if the transaction is on-chain, false otherwise.

pub async fn check_utxo_address_and_amount( &self, outpoint: &OutPoint, address: &ScriptBuf, amount_sats: Amount, ) -> Result<bool, BitcoinRPCError>

Checks if a transaction UTXO has expected address and amount.

Parameters

outpoint - The outpoint to check

address - Expected script pubkey

amount_sats - Expected amount in satoshis

Returns

bool: true if the UTXO has the expected address and amount, false otherwise.

pub async fn is_utxo_spent( &self, outpoint: &OutPoint, ) -> Result<bool, BitcoinRPCError>

Checks if an UTXO is spent.

Parameters

outpoint: The outpoint to check

Returns

bool: true if the UTXO is spent, false otherwise.

Errors

BitcoinRPCError: If the transaction is not confirmed or if there was an error retrieving the transaction output.

pub async fn mempool_size(&self) -> Result<usize, BitcoinRPCError>

Gets the number of transactions in the mempool.

Returns

usize: The number of transactions in the mempool.

pub async fn send_to_address( &self, address: &Address, amount_sats: Amount, ) -> Result<OutPoint, BitcoinRPCError>

Sends a specified amount of Bitcoins to the given address.

Parameters

address - The recipient address

amount_sats - The amount to send in satoshis

Returns

[OutPoint]: The outpoint (txid and vout) of the newly created output.

pub async fn get_txout_from_outpoint( &self, outpoint: &OutPoint, ) -> Result<TxOut, BitcoinRPCError>

Retrieves the transaction output for a given outpoint.

Arguments

outpoint - The outpoint (txid and vout) to retrieve

Returns

[TxOut]: The transaction output at the specified outpoint.

pub async fn bump_fee_with_fee_rate( &self, txid: Txid, fee_rate: FeeRate, ) -> Result<Txid, BitcoinRPCError>

Bumps the fee of a transaction to meet or exceed a target fee rate. Does nothing if the transaction is already confirmed. Returns the original txid if no bump was needed.

This function implements Replace-By-Fee (RBF) to increase the fee of an unconfirmed transaction. It works as follows:

If the transaction is already confirmed, returns Err(TransactionAlreadyInBlock)

If the current fee rate is already >= the requested fee rate, returns the original txid

Otherwise, increases the fee rate by adding the node’s incremental fee to the current fee rate, then bump_fees the transaction

Note: This function currently only supports fee payer TXs.

Arguments

txid - The transaction ID to bump

fee_rate - The target fee rate to achieve

Returns

[Txid]: The txid of the bumped transaction (which may be the same as the input txid if no bump was needed).

Errors

TransactionAlreadyInBlock - If the transaction is already confirmed

BumpFeeUTXOSpent - If the UTXO being spent by the transaction is already spent

BumpFeeError - For other errors with fee bumping

pub async fn clone_inner(&self) -> Result<Self, Error>

Creates a new instance of the ExtendedBitcoinRpc with a new client connection for cloning. This is needed when you need a separate connection to the Bitcoin RPC server.

Returns

ExtendedBitcoinRpc: A new instance of ExtendedBitcoinRpc with a new client connection.