clementine_core
Macro fetch_next_message_from_stream

macro_rules! fetch_next_message_from_stream {
    ($stream:expr, $field:ident) => { ... };
}

Fetches the next message from a stream which is unwrapped and encapsulated by a Result.

Parameters

stream: tonic::Streaming typed input stream

field: Input field ident (struct member) to look in the next message

Returns

A Result containing the next message. Will return an Err variant if stream has exhausted.