clementine_core

Module musig2

MuSig2

Helper functions for the MuSig2 signature scheme.

Structs

TAPROOT_TWEAK_TAGGED_HASH

Enums

Musig2Mode

Possible Musig2 modes.

Constants

TAPROOT_TWEAK_TAG_DIGEST 🔒

sha256(b“TapTweak“)

Traits

AggregateFromPublicKeys

Functions

aggregate_nonces

aggregate_partial_signatures

Aggregates the partial signatures into a single aggregated signature and verifies the aggregated signature. If PARTIAL_SIG_VERIFICATION is set to true, each partial signature will also be verified.

create_key_agg_cache 🔒

from_secp_kp

from_secp_pk

from_secp_sig

from_secp_xonly

nonce_pair

Generates a pair of nonces, one secret and one public. Be careful, DO NOT REUSE the same pair of nonces for multiple transactions. It will cause you to leak your secret key. For more information. See: https://medium.com/blockstream/musig-dn-schnorr-multisignatures-with-verifiably-deterministic-nonces-27424b5df9d6#e3b6.

partial_sign

to_secp_kp

to_secp_msg

to_secp_pk

to_secp_sk

Type Aliases

MuSigNoncePair