clementine_core
Module operator

Modules

states ๐

Structs

Operator

OperatorServer

Enums

RoundIndex

Round index is used to represent the round index safely. Collateral represents the collateral utxo. Round(index) represents the rounds of the bridge operators, index is 0-indexed. As a single u32, collateral is represented as 0 and rounds are represented starting from 1.

Type Aliases

PublicHash

SecretPreimage