clementine_core::operator
Struct Operator

pub struct Operator<C: CitreaClientT> {
    pub rpc: ExtendedRpc,
    pub db: Database,
    pub signer: Actor,
    pub config: BridgeConfig,
    pub collateral_funding_outpoint: OutPoint,
    pub(crate) reimburse_addr: Address,
    pub tx_sender: TxSenderClient,
    pub header_chain_prover: HeaderChainProver,
    pub citrea_client: C,
}

Fields
rpc: ExtendedRpcdb: Databasesigner: Actorconfig: BridgeConfigcollateral_funding_outpoint: OutPointreimburse_addr: Addresstx_sender: TxSenderClientheader_chain_prover: HeaderChainProvercitrea_client: C
Implementations

impl<C> Operator<C>
where C: CitreaClientT,

pub async fn create_assert_commitment_txs( &self, assert_data: TransactionRequestData, commit_data: Vec<Vec<Vec<u8>>>, ) -> Result<Vec<(TransactionType, Transaction)>, BridgeError>

Creates and signs all the assert commitment transactions for a single kickoff of an operator.

Arguments

assert_data - Data to identify the deposit and kickoff.

commit_data - BitVM assertions for the kickoff, for each assert tx.

Returns

A vector of tuples: 1. TransactionType: MiniAssert(idx) for idx’th assert commitment 2. Transaction: Signed assert commitment transaction

pub async fn create_latest_blockhash_tx( &self, assert_data: TransactionRequestData, block_hash: BlockHash, ) -> Result<(TransactionType, Transaction), BridgeError>

Creates and signs the latest blockhash transaction for a single kickoff of an operator.

Arguments

assert_data - Data to identify the deposit and kickoff.

block_hash - Block hash to commit using winternitz signatures.

Returns

A tuple of: 1. TransactionType: LatestBlockhash 2. Transaction: Signed latest blockhash transaction

impl<C> Operator<C>
where C: CitreaClientT,

pub async fn new(config: BridgeConfig) -> Result<Self, BridgeError>

Creates a new Operator.

pub async fn send_initial_round_tx( &self, round_tx: &Transaction, ) -> Result<(), BridgeError>

pub async fn get_params( &self, ) -> Result<(Receiver<PublicKey>, Receiver<Signature>), BridgeError>

Returns an operator’s winternitz public keys and challenge ackpreimages & hashes.

Returns

[mpsc::Receiver]: A [tokio] data channel with a type of [winternitz::PublicKey] and size of operator’s winternitz public keys count

[mpsc::Receiver]: A [tokio] data channel with a type of PublicHash and size of operator’s challenge ack preimages & hashes count

pub async fn deposit_sign( &self, deposit_data: DepositData, ) -> Result<Receiver<Signature>, BridgeError>

pub async fn track_rounds(&self) -> Result<(), BridgeError>

Creates the round state machine by adding a system event to the database

fn is_profitable( input_amount: Amount, withdrawal_amount: Amount, bridge_amount_sats: Amount, operator_withdrawal_fee_sats: Amount, ) -> bool

Checks if the withdrawal amount is within the acceptable range.

pub async fn withdraw( &self, withdrawal_index: u32, in_signature: Signature, in_outpoint: OutPoint, out_script_pubkey: ScriptBuf, out_amount: Amount, ) -> Result<(), BridgeError>

Prepares a withdrawal by:

Checking if the withdrawal has been made on Citrea

Verifying the given signature

Checking if the withdrawal is profitable or not

Funding the witdhrawal transaction using TxSender RBF option

Parameters

withdrawal_idx: Citrea withdrawal UTXO index

in_signature: User’s signature that is going to be used for signing withdrawal transaction input

in_outpoint: User’s input for the payout transaction

out_script_pubkey: User’s script pubkey which will be used in the payout transaction’s output

out_amount: Payout transaction output’s value

Returns

Ok(()) if the withdrawal checks are successful and a payout transaction is added to the TxSender

Err(BridgeError) if the withdrawal checks fail

pub fn generate_assert_winternitz_pubkeys( &self, deposit_outpoint: OutPoint, ) -> Result<Vec<PublicKey>, BridgeError>

Generates Winternitz public keys for every BitVM assert tx for a deposit.

Returns

Vec<Vec<winternitz::PublicKey>>: Winternitz public keys for watchtower index row and BitVM assert tx index column.

pub fn generate_kickoff_winternitz_pubkeys( &self, ) -> Result<Vec<PublicKey>, BridgeError>

Generates Winternitz public keys for every blockhash commit to be used in kickoff utxos. Unique for each kickoff utxo of operator.

Returns

Vec<Vec<winternitz::PublicKey>>: Winternitz public keys for round_index row and kickoff_idx column.

pub fn generate_unspent_kickoff_sigs( &self, kickoff_wpks: &KickoffWinternitzKeys, ) -> Result<Vec<Signature>, BridgeError>

pub fn generate_challenge_ack_preimages_and_hashes( &self, deposit_data: &DepositData, ) -> Result<Vec<PublicHash>, BridgeError>

pub async fn handle_finalized_payout<'a>( &'a self, dbtx: DatabaseTransaction<'a, '_>, deposit_outpoint: OutPoint, payout_tx_blockhash: BlockHash, ) -> Result<Txid, BridgeError>

async fn start_first_round( &self, dbtx: DatabaseTransaction<'_, '_>, kickoff_wpks: KickoffWinternitzKeys, ) -> Result<(), BridgeError>

pub async fn end_round<'a>( &'a self, dbtx: DatabaseTransaction<'a, '_>, ) -> Result<(), BridgeError>

async fn send_asserts( &self, kickoff_data: KickoffData, deposit_data: DepositData, watchtower_challenges: HashMap<usize, Transaction>, _payout_blockhash: Witness, latest_blockhash: Witness, ) -> Result<(), BridgeError>

fn data(&self) -> OperatorData

async fn send_latest_blockhash( &self, kickoff_data: KickoffData, deposit_data: DepositData, latest_blockhash: BlockHash, ) -> Result<(), BridgeError>