clementine_core::rpc

Module aggregator

Structs

AggNonceQueueItem 🔒

FinalSigQueueItem 🔒

Enums

AggregatorError

Functions

collect_and_call 🔒

Use items collected from the broadcast receiver for an async function call.

create_nonce_streams 🔒

Creates a stream of nonces from verifiers. This will automatically get the first response from the verifiers.

get_next_pub_nonces 🔒

nonce_aggregator 🔒

For each expected sighash, we collect a batch of public nonces from all verifiers. We aggregate and send aggregated nonce and all public nonces (needed for partial signature verification) to the agg_nonce_sender. Then repeat for the next sighash.

nonce_distributor 🔒

Reroutes aggregated nonces and public nonces for each aggregated nonce to the signature aggregator.

signature_aggregator 🔒

Collects partial signatures and corresponding public nonces from given stream and aggregates them. Each partial signature will also be verified if PARTIAL_SIG_VERIFICATION is set to true.

signature_distributor 🔒

Reroutes aggregated signatures to the caller. Also sends 2 aggregated nonces to the verifiers.