clementine_core::rpc
Module clementine

Modules

clementine_aggregator_client

Generated client implementations.

clementine_aggregator_server

Generated server implementations.

clementine_operator_client

Generated client implementations.

clementine_operator_server

Generated server implementations.

clementine_verifier_client

Generated client implementations.

clementine_verifier_server

Generated server implementations.

deposit

Nested message and enum types in Deposit.

entity_status_with_id

Nested message and enum types in EntityStatusWithId.

grpc_transaction_id

Nested message and enum types in GrpcTransactionId.

nonce_gen_response

Nested message and enum types in NonceGenResponse.

operator_params

Nested message and enum types in OperatorParams.

tagged_signature

Nested message and enum types in TaggedSignature.

verifier_deposit_finalize_params

Nested message and enum types in VerifierDepositFinalizeParams.

verifier_deposit_sign_params

Nested message and enum types in VerifierDepositSignParams.

Structs

Actors

AggregatorWithdrawResponse

AggregatorWithdrawalInput

Input of the aggregator’s withdraw function. It contains the withdrawal params along with the verification signature that signs the withdrawal params. It also contains the operator’s xonly public keys that the withdrawal request should be sent to. If the list is empty, the withdrawal will be sent to all operators.

BaseDeposit

A new original deposit request’s details.

ChallengeAckDigest

Deposit

DepositParams

DepositSignSession

Includes the deposit params and the nonce gen initial responses (pubkeys and their signatures from all verifiers)

DepositSignatures

Empty

EntityError

EntityId

EntityStatus

EntityStatusWithId

EntityStatuses

FinalizedPayoutParams

GetEmergencyStopTxRequest

GetEmergencyStopTxResponse

GetEntityStatusesRequest

GrpcTransactionId

KickoffId

NofnResponse

NonceGenFirstResponse

NonceGenRequest

NonceGenResponse

NormalSignatureId

NumberedSignatureId

NumberedTransactionId

OperatorConfig

OperatorKeys

OperatorKeysWithDeposit

OperatorParams

OptimisticPayoutParams

OptimisticWithdrawParams

Outpoint

PartialSig

RawSignedTx

RawSignedTxs

RawTxWithRbfInfo

RbfSigningInfoRpc

ReplacementDeposit

SchnorrSig

SecurityCouncil

SendMoveTxRequest

SendTxRequest

SignedTxWithType

SignedTxsWithType

StoppedTasks

TaggedSignature

A tagged signature struct that identifies the transaction-input that the signature is for. The id is left as NotStored for signatures that are created on the fly by the operator (they’re also not stored).

TransactionRequest

TxDebugFeePayerUtxo

TxDebugInfo

TxDebugRequest

TxDebugSubmissionError

TxMetadata

Txid

VergenResponse

VerifierDepositFinalizeParams

VerifierDepositFinalizeResponse

VerifierDepositSignParams

VerifierParams

VerifierPublicKeys

WinternitzPubkey

WithdrawParams

WithdrawParamsWithSig

XOnlyPublicKeyRpc

XOnlyPublicKeys

Enums

EntityType

FeeType

NormalSignatureKind

NormalTransactionId

NumberedSignatureKind

Signatures that are needed multiple times per an operators kickoff. Some watchtower sigs are needed once per watchtower. Asserts are needed multiple times

NumberedTransactionType