clementine_core::rpc

Module ecdsa_verification_sig

ECDSA Verification Signature

This module contains the ECDSA verification signature for the Clementine protocol. It is for additional verification that the request for optimistic payout and withdrawal is coming from the aggregator, which is the owner of the address in operator/verifiers config.

The address who signed the signature is retrieved by calculating the EIP-712 hash of the withdrawal params. The address is then compared to the address in the config.

Structs

OperatorWithdrawalMessage

OptimisticPayoutMessage

Statics

CLEMENTINE_EIP712_DOMAIN

Traits

WithdrawalMessage

Functions

recover_address_from_ecdsa_signature

Recover the address from the signature EIP712 hash is calculated from optimistic payout params Signature is the signature of the eip712 hash