clementine_core::states
Module context

Structs

StateContext

Enums

Duty

Duties are notifications that are sent to the owner (verifier or operator) of the state machine to notify them on changes to the current contract state that require action. Note that for all kickoff state duties, they are only sent if withdrawal process is still going on, meaning the burn connector and kickoff finalizer is still on-chain/unspent.

DutyResult

Result of handling a duty

Traits

Owner

Owner trait with async handling and tx handler creation