clementine_core::states

Module kickoff

Structs

KickoffStateMachine

State machine for tracking a single kickoff process in the protocol.

Enums

KickoffEvent

Events that can be dispatched to the kickoff state machine These event either trigger state transitions or trigger actions of the owner

State

Superstate