clementine_core::states

Trait ContextProcessor

pub(crate) trait ContextProcessor<T: Owner, M: IntoStateMachine> {
    // Required method
    fn process_with_ctx(
        self,
        block: &StateContext<T>,
    ) -> ContextProcessResult<T, M, impl Future<Output = (InitializedStateMachine<M>, StateContext<T>)> + Send>;
}

Utility trait to make processing generic

Required Methods

fn process_with_ctx( self, block: &StateContext<T>, ) -> ContextProcessResult<T, M, impl Future<Output = (InitializedStateMachine<M>, StateContext<T>)> + Send>

Processes the machine with the given state context (which contains the block cache) If the machine is unchanged, it is returned as is. Otherwise, the machine is processed and the result is returned as a future that processes the new events.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.