clementine_core

Module task

Modules

aggregator_metric_publisher

entity_metric_publisher

manager

payout_checker

status_monitor

Structs

BufferedErrors

CancelableLoop

CancelableTask

A task that can be canceled via a oneshot channel

IgnoreError

A task that ignores errors from the inner task and returns a default value.

Map

WithDelay

A task that adds a certain delay after the inner task has run to reduce polling frequency. When inner returns false, the delay is applied.

Enums

CancelableResult

TaskVariant

The variant of the task, used for identifying the task in the status monitor Create a new enum variant for each task that you want to track in the status monitor BackgroundTaskManager will use TaskVariant to identify the tasks, to not start the same task twice.

Traits

IntoTask

A trait for objects that can be converted into a Task

Task

Task trait defining the core behavior for cancelable background tasks

TaskExt