clementine_core

Module tx_sender

Structs

ActivatedWithOutpoint

ActivatedWithTxid

RbfSigningInfo

Information to re-sign an RBF transaction. Specifically the merkle root of the taproot to keyspend with and the output index of the utxo to be re-signed.

TxMetadata

TxSender

Manages the process of sending Bitcoin transactions, including handling fee bumping strategies like Replace-By-Fee (RBF) and Child-Pays-For-Parent (CPFP).

TxSenderClient

TxSenderTask

Enums

FeePayingType

Specifies the fee bumping strategy used for a transaction.

SendTxError