clementine_core::builder
Module transaction

Transaction Builder

Transaction builder provides useful functions for building typical Bitcoin transactions.

Modules

deposit_signature_owner

input

output

sign

Structs

Actors

BaseDepositData

Type to uniquely identify a deposit.

ContractContext

Context for a single operator and round, and optionally a single deposit

DepositData

DepositInfo

KickoffData

KickoffWinternitzKeys

Helper struct to get specific kickoff winternitz keys for a sequential collateral tx

OperatorData

ReimburseDbCache

Struct to retrieve and cache data from DB for creating TxHandlers on demand It can only store information for one deposit and operator pair.

ReplacementDepositData

SecurityCouncil

Signed

TxHandler

TxHandlerBuilder

TxHandlerCache

Struct to store common txhandlers for kickoffs

Unsigned

Enums

DepositType

TransactionType

Types of all transactions that can be created. Some transactions have an (usize) to as they are created multiple times per kickoff.

TxError

Constants

DEFAULT_SEQUENCE

Traits

State

Functions

anchor_output

Creates a P2A output for CPFP.

anyone_can_spend_txout

Creates a P2WSH output that anyone can spend. TODO: We will not need this in the future.

combine_emergency_stop_txhandler

We assume that the vector of (Txid, Transaction) includes input-output pairs which are signed using Sighash Single | AnyoneCanPay. This function will combine the inputs and outputs of the transactions into a single transaction. Beware, this may be dangerous, as there are no checks.

create_burn_unused_kickoff_connectors_txhandler

create_disprove_taproot_output

create_emergency_stop_txhandler

Creates a TxHandlerBuilder for the emergency_stop_tx. This transaction will move the funds to a the address controlled by the security council from the move to vault txout This transaction will be used to stop malicious activities if there is a security issue.

create_move_to_vault_txhandler

Creates a TxHandler for the move_to_vault_tx. This transaction will move the funds to a NofN address from the deposit intent address, after all the signature collection operations are done.

create_optimistic_payout_txhandler

Creates a TxHandler for the optimistic_payout_tx. This transaction will be signed by all verifiers that participated in the corresponding deposit to directly payout without any kickoff.

create_payout_txhandler

Creates a TxHandler for the payout_tx. This transaction will be sent by the operator for withdrawals.

create_replacement_deposit_txhandler

Creates a TxHandlerBuilder for the move_to_vault_tx. This transaction will move the funds to a NofN address from the deposit intent address, after all the signature collection operations are done.

create_round_nth_txhandler

Creates the nth (0-indexed) round_txhandler and reimburse_generator_txhandler pair for a specific operator.

create_round_txhandlers

Function to create the round txhandler and ready to reimburse txhandler for a specific operator and round index

create_taproot_output_with_hidden_node

Helper function to create a taproot output that combines a script and a root hash

create_txhandlers

op_return_txout

Creates a OP_RETURN output.

remove_txhandler_from_map

Type Aliases

SighashCalculator