clementine_core::config
Module protocol

Structs

ProtocolParamset

Protocol parameters that affect the transactions in the contract (which also change the pre-calculated txids and sighashes).

Enums

ProtocolParamsetName

A pre-defined paramset name that can be converted into a ProtocolParamset reference.

Constants

BLOCKS_PER_DAY

BLOCKS_PER_HOUR

BLOCKS_PER_WEEK

MAINNET_BRIDGE_CIRCUIT_CONSTANT

REGTEST_BRIDGE_CIRCUIT_CONSTANT

REGTEST_PARAMSET

REGTEST_TEST_BRIDGE_CIRCUIT_CONSTANT

SIGNET_BRIDGE_CIRCUIT_CONSTANT

TESTNET4_BRIDGE_CIRCUIT_CONSTANT

TESTNET4_TEST_PARAMSET

WINTERNITZ_LOG_D

This is the log_d used across the codebase.

Functions

convert_hex_string_to_bytes ๐