clementine_core
Module deposit

This module defines the data structures related to Citrea deposits in the Clementine bridge.

It includes structures for representing deposit data, actors involved (verifiers, watchtowers, operators), and security council configurations. The module also provides functionality for managing different types of deposits (base and replacement) and deriving the necessary scripts these deposits must have.

Structs

Actors

Data structure to represent the actors public keys that participate in the deposit.

BaseDepositData

Data structure to represent the data for a base deposit. These kinds of deposits are created by users.

DepositData

Data structure to represent a deposit.

DepositInfo

Data structure to represent the deposit outpoint and type.

KickoffData

Data structure to represent a single kickoff utxo in an operators round tx.

OperatorData

Data structure to represent the data for an operator. These data is used in the tx creation so any deviation will change the tx’s created by the bridge.

ReplacementDepositData

Data structure to represent the data for a replacement deposit. These kinds of deposits are created by the bridge, using security council to unlock the previous deposit and move the funds to create a new deposit. Verifiers will sign the new deposit again.

SecurityCouncil

Data structure to represent the security council that can unlock the deposit using an m-of-n multisig to create a replacement deposit.

Enums

DepositType

Type to represent the type of deposit, and related specific data for each type..