clementine_core::rpc::clementine::clementine_verifier_server

Trait ClementineVerifier

pub trait ClementineVerifier:
    Send
    + Sync
    + 'static {
    type NonceGenStream: Stream<Item = Result<NonceGenResponse, Status>> + Send + 'static;
    type DepositSignStream: Stream<Item = Result<PartialSig, Status>> + Send + 'static;

    // Required methods
    fn get_params<'life0, 'async_trait>(
        &'life0 self,
        request: Request<Empty>,
    ) -> Pin<Box<dyn Future<Output = Result<Response<VerifierParams>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn get_compatibility_params<'life0, 'async_trait>(
        &'life0 self,
        request: Request<Empty>,
    ) -> Pin<Box<dyn Future<Output = Result<Response<CompatibilityParamsRpc>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn set_operator<'life0, 'async_trait>(
        &'life0 self,
        request: Request<Streaming<OperatorParams>>,
    ) -> Pin<Box<dyn Future<Output = Result<Response<Empty>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn set_operator_keys<'life0, 'async_trait>(
        &'life0 self,
        request: Request<OperatorKeysWithDeposit>,
    ) -> Pin<Box<dyn Future<Output = Result<Response<Empty>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn nonce_gen<'life0, 'async_trait>(
        &'life0 self,
        request: Request<NonceGenRequest>,
    ) -> Pin<Box<dyn Future<Output = Result<Response<Self::NonceGenStream>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn deposit_sign<'life0, 'async_trait>(
        &'life0 self,
        request: Request<Streaming<VerifierDepositSignParams>>,
    ) -> Pin<Box<dyn Future<Output = Result<Response<Self::DepositSignStream>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn optimistic_payout_sign<'life0, 'async_trait>(
        &'life0 self,
        request: Request<OptimisticPayoutParams>,
    ) -> Pin<Box<dyn Future<Output = Result<Response<PartialSig>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn deposit_finalize<'life0, 'async_trait>(
        &'life0 self,
        request: Request<Streaming<VerifierDepositFinalizeParams>>,
    ) -> Pin<Box<dyn Future<Output = Result<Response<VerifierDepositFinalizeResponse>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn debug_tx<'life0, 'async_trait>(
        &'life0 self,
        request: Request<TxDebugRequest>,
    ) -> Pin<Box<dyn Future<Output = Result<Response<TxDebugInfo>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn restart_background_tasks<'life0, 'async_trait>(
        &'life0 self,
        request: Request<Empty>,
    ) -> Pin<Box<dyn Future<Output = Result<Response<Empty>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn internal_handle_kickoff<'life0, 'async_trait>(
        &'life0 self,
        request: Request<Txid>,
    ) -> Pin<Box<dyn Future<Output = Result<Response<Empty>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn get_current_status<'life0, 'async_trait>(
        &'life0 self,
        request: Request<Empty>,
    ) -> Pin<Box<dyn Future<Output = Result<Response<EntityStatus>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn internal_create_signed_txs<'life0, 'async_trait>(
        &'life0 self,
        request: Request<TransactionRequest>,
    ) -> Pin<Box<dyn Future<Output = Result<Response<SignedTxsWithType>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn internal_create_watchtower_challenge<'life0, 'async_trait>(
        &'life0 self,
        request: Request<TransactionRequest>,
    ) -> Pin<Box<dyn Future<Output = Result<Response<RawTxWithRbfInfo>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn vergen<'life0, 'async_trait>(
        &'life0 self,
        request: Request<Empty>,
    ) -> Pin<Box<dyn Future<Output = Result<Response<VergenResponse>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;

}

Generated trait containing gRPC methods that should be implemented for use with ClementineVerifierServer.

Required Associated Types

type NonceGenStream: Stream<Item = Result<NonceGenResponse, Status>> + Send + 'static

Server streaming response type for the NonceGen method.

type DepositSignStream: Stream<Item = Result<PartialSig, Status>> + Send + 'static

Server streaming response type for the DepositSign method.

Required Methods

fn get_params<'life0, 'async_trait>( &'life0 self, request: Request<Empty>, ) -> Pin<Box<dyn Future<Output = Result<Response<VerifierParams>, Status>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Returns verifiers’ metadata. Needs to be called once per setup.

Used by aggregator inside setup to let all verifiers know all other verifier pks

fn get_compatibility_params<'life0, 'async_trait>( &'life0 self, request: Request<Empty>, ) -> Pin<Box<dyn Future<Output = Result<Response<CompatibilityParamsRpc>, Status>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Returns the protocol params that can affect the transactions in the contract, syncing with citrea and version number for checking compatibility

fn set_operator<'life0, 'async_trait>( &'life0 self, request: Request<Streaming<OperatorParams>>, ) -> Pin<Box<dyn Future<Output = Result<Response<Empty>, Status>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Saves an operator.

Used by aggregator inside setup to let all verifiers know all other operator pks

fn set_operator_keys<'life0, 'async_trait>( &'life0 self, request: Request<OperatorKeysWithDeposit>, ) -> Pin<Box<dyn Future<Output = Result<Response<Empty>, Status>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Sets the operator’s winternitz keys and challenge ACK hashes and saves them into the db.

Used by aggregator inside new_deposit to let all verifiers know all other operators’ deposit information

fn nonce_gen<'life0, 'async_trait>( &'life0 self, request: Request<NonceGenRequest>, ) -> Pin<Box<dyn Future<Output = Result<Response<Self::NonceGenStream>, Status>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Generates nonces for a deposit.

Returns

Nonce metadata followed by nonces.

Used by aggregator inside new_deposit

fn deposit_sign<'life0, 'async_trait>( &'life0 self, request: Request<Streaming<VerifierDepositSignParams>>, ) -> Pin<Box<dyn Future<Output = Result<Response<Self::DepositSignStream>, Status>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Signs deposit with given aggNonces and verifier’s secNonce using nonce_id.

Used by aggregator inside new_deposit

fn optimistic_payout_sign<'life0, 'async_trait>( &'life0 self, request: Request<OptimisticPayoutParams>, ) -> Pin<Box<dyn Future<Output = Result<Response<PartialSig>, Status>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Signs the optimistic payout tx with given aggNonce and withdrawal info.

fn deposit_finalize<'life0, 'async_trait>( &'life0 self, request: Request<Streaming<VerifierDepositFinalizeParams>>, ) -> Pin<Box<dyn Future<Output = Result<Response<VerifierDepositFinalizeResponse>, Status>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Verifies every signature and signs move_tx.

Used by aggregator inside new_deposit

fn debug_tx<'life0, 'async_trait>( &'life0 self, request: Request<TxDebugRequest>, ) -> Pin<Box<dyn Future<Output = Result<Response<TxDebugInfo>, Status>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Debug a transaction by retrieving its current state and history

fn restart_background_tasks<'life0, 'async_trait>( &'life0 self, request: Request<Empty>, ) -> Pin<Box<dyn Future<Output = Result<Response<Empty>, Status>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Restarts the background tasks for the verifier.

fn internal_handle_kickoff<'life0, 'async_trait>( &'life0 self, request: Request<Txid>, ) -> Pin<Box<dyn Future<Output = Result<Response<Empty>, Status>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Checks if the kickoff tx is malicious and if so, try to send all necessary txs to punish the operator

fn get_current_status<'life0, 'async_trait>( &'life0 self, request: Request<Empty>, ) -> Pin<Box<dyn Future<Output = Result<Response<EntityStatus>, Status>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Returns the current status of tasks running on the verifier and their last synced heights.

fn internal_create_signed_txs<'life0, 'async_trait>( &'life0 self, request: Request<TransactionRequest>, ) -> Pin<Box<dyn Future<Output = Result<Response<SignedTxsWithType>, Status>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Signs all tx’s it can according to given transaction type (use it with AllNeededForDeposit to get almost all tx’s)

Creates the transactions denoted by the deposit and operator_idx, round_idx, and kickoff_idx.

It will create the transaction and sign it with the operator’s private key and/or saved nofn signatures.

Parameters

deposit_params: User’s deposit information

transaction_type: Requested Transaction type

kickoff_id: Operator’s kickoff ID

Returns

Raw signed transactions that the entity can sign (no asserts and watchtower challenge)

fn internal_create_watchtower_challenge<'life0, 'async_trait>( &'life0 self, request: Request<TransactionRequest>, ) -> Pin<Box<dyn Future<Output = Result<Response<RawTxWithRbfInfo>, Status>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Signs the verifiers own watchtower challenge tx in the corresponding kickoff and returns the signed raw tx

fn vergen<'life0, 'async_trait>( &'life0 self, request: Request<Empty>, ) -> Pin<Box<dyn Future<Output = Result<Response<VergenResponse>, Status>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,