clementine_core
Module tx_sender

Modules

client 🔒

cpfp 🔒

rbf 🔒

task 🔒

Macros

log_error_for_tx 🔒

Structs

ActivatedWithOutpoint

ActivatedWithTxid

TxSender

Manages the process of sending Bitcoin transactions, including handling fee bumping strategies like Replace-By-Fee (RBF) and Child-Pays-For-Parent (CPFP).

TxSenderClient

TxSenderTask

Enums

SendTxError

Functions

get_fee_rate_from_mempool_space 🔒

Fetches the current recommended fee rate from RPC provider. Currently only supports Mempool Space API. This function is used to get the fee rate in sat/vkb (satoshis per kilovbyte). See Mempool Space API for more details.

Type Aliases

Result 🔒