clementine_core

Module tx_sender

Transaction Sender

Transaction sender is responsible for sending Bitcoin transactions, bumping fees and making sure that transactions are finalized until the deadline. It can utilize Child-Pays-For-Parent (CPFP) and Replace-By-Fee (RBF) strategies for sending transactions.

Sending transactions is done by the TxSenderClient, which is a client that puts transactions into the sending queue and the TxSenderTask is responsible for processing this queue and sending them.

Debugging Transaction Sender

There are several database tables that saves the transaction states. Please look for [core/src/database/tx_sender.rs] for more information.

Modules

client 🔒

Transaction Sender Client

cpfp 🔒

Child Pays For Parent (CPFP) Support For Transaction Sender

nonstandard 🔒

TxSender for nonstandard transactions.

rbf 🔒

task 🔒

Transaction Sender Task

Macros

log_error_for_tx 🔒

Structs

ActivatedWithOutpoint

ActivatedWithTxid

TxSender

Manages the process of sending Bitcoin transactions, including handling fee bumping strategies like Replace-By-Fee (RBF) and Child-Pays-For-Parent (CPFP).

TxSenderClient

TxSenderTask

Enums

SendTxError

Type Aliases

Result 🔒