clementine_core::utils
Trait NamedEntity

pub trait NamedEntity:
    Sync
    + Send
    + 'static {
    const ENTITY_NAME: &'static str;
    const TX_SENDER_CONSUMER_ID: &'static str;
    const FINALIZED_BLOCK_CONSUMER_ID: &'static str;
}

A trait for entities that have a name, operator, verifier, etc. Used to distinguish between state machines with different owners in the database, and to provide a human-readable name for the entity for task names.

Required Associated Constants

const ENTITY_NAME: &'static str

A string identifier for this owner type used to distinguish between state machines with different owners in the database.

Example

“operator”, “verifier”, “user”

const TX_SENDER_CONSUMER_ID: &'static str

Consumer ID for the tx sender task.

const FINALIZED_BLOCK_CONSUMER_ID: &'static str

Consumer ID for the finalized block task.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.